fix #1 - UTF-8 decode problems when using en_US.
Apparently the en_US dictionary in ubuntu cannot be decoded with utf-8,
but works fine with iso-8859-1. As a workaround this adds another
configuration and command line option: encoding. If left empty it
autodetects from the environment, otherwise it should be an encoding
definition that python should use to read the dictionary.